requirements.txt
aniso8601==7.0.0
Django==2.2.10
django-graphiql==0.4.4
django-rest-framework==0.1.0
djangorestframework==3.11.0
graphene==2.1.8
graphene-django==2.12.1
graphql-core==2.3.2
graphql-relay==2.0.1
pkg-resources==0.0.0
promise==2.3
pytz==2020.1
Rx==1.6.1
singledispatch==3.4.0.3
six==1.15.0
sqlparse==0.3.1
Unidecode==1.1.1
